move gps toggle to App component
