+++ app/components/App.js
@@ -1,6 +1,7 @@
 import React from 'react';
-import { StatusBar, View } from 'react-native';
+import { AppState, StatusBar, View } from 'react-native';
 import { StackNavigator } from 'react-navigation';
+import * as sc from 'react-native-spatialconnect';
 import { blue, darkBlue } from '../styles';
 import FormList from './FormList';
 import Form from './Form';
@@ -28,11 +29,50 @@ const AppStack = StackNavigator(
   }
 );
 
-const App = () => (
-  <View style={{ flex: 1 }}>
-    <StatusBar backgroundColor={darkBlue} barStyle="light-content" />
-    <AppStack />
-  </View>
-);
+class App extends React.Component {
+  componentDidMount() {
+    sc.addConfigFilepath('layers.scfg');
+    sc.startAllServices();
+    AppState.addEventListener('change', this._handleAppStateChange);
+  }
+
+  componentWillUnmount() {
+    AppState.removeEventListener('change', this._handleAppStateChange);
+  }
+
+  _handleAppStateChange = nextAppState => {
+    if (nextAppState === 'active') {
+      if (Platform.OS === 'android' && Platform.Version >= 23) {
+        try {
+          const granted = PermissionsAndroid.request(
+            PermissionsAndroid.PERMISSIONS.ACCESS_FINE_LOCATION,
+            {
+              title: 'GPS permission',
+              message: 'UNCollector needs access to your GPS',
+            }
+          );
+          if (granted) {
+            sc.enableGPS();
+          }
+        } catch (err) {
+          console.warn(err);
+        }
+      } else {
+        sc.enableGPS();
+      }
+    } else {
+      sc.disableGPS();
+    }
+  };
+
+  render() {
+    return (
+      <View style={{ flex: 1 }}>
+        <StatusBar backgroundColor={darkBlue} barStyle="light-content" />
+        <AppStack />
+      </View>
+    );
+  }
+}
 
 export default App;
